Major changes to CRUD logic and meeting list styling

+++ src/components/Book/BookForm.tsx
@@ -1,5 +1,7 @@
 import {
+  Button,
   Dialog,
+  DialogActions,
   DialogTitle,
   FormControl,
   InputLabel,
@@ -9,7 +11,7 @@ import {
 } from '@mui/material';
 import { isBefore } from 'date-fns';
 import { DocumentData, deleteDoc, doc, updateDoc } from 'firebase/firestore';
-import { useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { db, firestore } from '../../firestore';
 import { FirestoreMeeting, MeetingInfo } from '../../pages';
 import { FirestoreBook, ReadStatus } from '../../pages/Books/Books';
@@ -42,17 +44,14 @@ export const BookForm = ({
   open,
   onClose,
 }: BookProps) => {
-  const [selectedReadStatus, setSelectedReadStatus] = useState<ReadStatus>();
+  const [selectedReadStatus, setSelectedReadStatus] = useState<
+    ReadStatus | string
+  >('');
   const [isOpen, setIsOpen] = useState<boolean>(false);
   const [meetings, setMeetings] = useState<FirestoreMeeting[]>([]);
   const [selectedMeeting, setSelectedMeeting] = useState<string | undefined>(
     ''
   );
-  const [showMeetingSelector, setShowMeetingSelector] =
-    useState<boolean>(false);
-
-  // const [addMeetingFromBook, setAddMeetingFromBook] =
-  //   useState<SelectChangeEvent<string> | null>(null);
 
   const booksRef = firestore.collection('books');
 
@@ -71,53 +70,52 @@ export const BookForm = ({
   }, [open]);
 
   useEffect(() => {
-    setSelectedReadStatus(readStatus);
+    setSelectedReadStatus(readStatus as ReadStatus);
     console.log({ volumeInfo, readStatus, id, scheduledMeeting });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [readStatus]);
 
   useEffect(() => {
     if (scheduledMeeting) {
-      setShowMeetingSelector(true);
       setSelectedMeeting(scheduledMeeting);
     }
   }, [scheduledMeeting]);
 
   useEffect(() => {
-    if (selectedMeeting) {
-      addBookToMeeting();
+    // Every time selectedReadStatus or SelectedMeeting change (and readStatus is not the initial state change), the book status should change
+    if (
+      (selectedReadStatus || selectedMeeting) &&
+      (selectedReadStatus !== readStatus ||
+        selectedMeeting !== scheduledMeeting)
+    ) {
+      handleBookStatus();
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selectedMeeting]);
+  }, [selectedReadStatus, selectedMeeting]);
 
   const addToShelf = async () => {
     // If the book does not exist on the shelf, add it
+    const date = new Date();
+    console.log(selectedReadStatus);
     if (
       !docId ||
       (!books.some((bookItem) => bookItem.data.id === id) && selectedReadStatus)
     ) {
-      const date = new Date();
       await booksRef.add({
         volumeInfo,
         id,
         addedDate: date.toLocaleDateString(),
         readStatus: selectedReadStatus,
-        scheduledMeeting: selectedMeeting ? selectedMeeting : '',
-      });
-      console.log('added to shelf', {
-        volumeInfo,
-        id,
-        addedDate: date.toLocaleDateString(),
-        readStatus: selectedReadStatus,
-        scheduledMeeting: selectedMeeting ? selectedMeeting : '',
+        scheduledMeeting: selectedMeeting,
       });
       // If the book already exists, update its status
-    } else if (docId) {
+    } else if (docId && selectedReadStatus) {
       const bookDocRef = doc(db, 'books', docId);
       try {
         await updateDoc(bookDocRef, {
           readStatus: selectedReadStatus,
-          scheduledMeeting: selectedMeeting ? selectedMeeting : '',
+          scheduledMeeting: selectedMeeting,
+          addedDate: date.toLocaleDateString(),
         });
       } catch (err) {
         alert(err);
@@ -125,87 +123,49 @@ export const BookForm = ({
     }
   };
 
-  const handleBookStatus = async (e: SelectChangeEvent) => {
-    setShowMeetingSelector(false); // Deactivate meeting selector by default
-    const selectedStatus = e.target.value as ReadStatus;
-    setSelectedReadStatus(selectedStatus);
-    switch (selectedStatus) {
+  const handleBookStatus = async () => {
+    if (!selectedReadStatus) {
+      return;
+    }
+    switch (selectedReadStatus) {
       case 'read':
       case 'reading':
-        setShowMeetingSelector(true);
+        if (selectedMeeting) {
+          addToShelf();
+          onClose();
+        }
         break;
       case 'candidate':
+        if (selectedMeeting) {
+          setSelectedMeeting('');
+          break;
+        }
         addToShelf();
         onClose();
         break;
-      default:
+      case 'unread':
         // If the book status is changed to "Unread", remove it.
         if (docId && books.some((bookItem) => bookItem.data.id === id)) {
           const bookDocRef = doc(db, 'books', docId);
           try {
             await deleteDoc(bookDocRef);
-            const meetingsWithBook = meetings.filter((meeting) =>
-              meeting.data.books?.some((book) => book.data.id === id)
-            );
-            // Remove the book from any meetings that is it currently on
-            meetingsWithBook.forEach(async (meeting) => {
-              const meetingDocRef = doc(db, 'meetings', meeting.docId);
-              const modifiedBookArray = meeting?.data?.books?.filter(
-                (book) => book.data.id !== id
-              );
-              await updateDoc(meetingDocRef, {
-                books: modifiedBookArray,
-              });
-            });
           } catch (err) {
             alert(err);
           }
         }
         onClose();
         break;
+      default: {
+        alert('No status was provided');
+      }
     }
   };
 
-  const addBookToMeeting = async () => {
-    const currentMeeting = meetings.find(
-      (meeting) => meeting.docId === selectedMeeting
-    );
-    const bookAlreadyAddedToMeeting = currentMeeting?.data?.books?.some(
-      (book) => book.data.id === id
-    );
-
-    if (!bookAlreadyAddedToMeeting && currentMeeting) {
-      const booksArray = (
-        currentMeeting?.data?.books?.length ? currentMeeting?.data?.books : []
-      ) as FirestoreBook[];
-      booksArray.push({
-        data: {
-          id,
-          volumeInfo,
-          readStatus: selectedReadStatus,
-          scheduledMeeting: selectedMeeting,
-        },
-      });
-
-      if (booksArray?.length && currentMeeting) {
-        currentMeeting.data.books = booksArray;
-        console.log('is this on?');
-      }
-    } else if (bookAlreadyAddedToMeeting) {
-      return alert(
-        `"${volumeInfo?.title}" is already scheduled for ${currentMeeting?.data.date}`
-      );
-    }
-    if (currentMeeting?.data.date && selectedMeeting) {
-      const addedDate = new Date();
-      const meetingDocRef = doc(db, 'meetings', selectedMeeting);
-      try {
-        addToShelf();
-        await updateDoc(meetingDocRef, { ...currentMeeting.data, addedDate });
-      } catch (err) {
-        alert(err);
-      }
-    }
+  const handleStatusSelect = (e: SelectChangeEvent) => {
+    // Reset meeting status.
+    setSelectedMeeting('');
+    // Set new read status
+    setSelectedReadStatus(e.target.value as ReadStatus);
   };
 
   return (
@@ -231,6 +191,7 @@ export const BookForm = ({
         </StyledBookBanner>
 
         {/* Select status form */}
+        {/* How to get the form to update properly when all values are present? */}
         <StyledModalBookForm>
           <StyledBookStatus>
             <FormControl fullWidth>
@@ -240,7 +201,7 @@ export const BookForm = ({
                 id="status-select"
                 value={selectedReadStatus}
                 label="Status"
-                onChange={(e) => handleBookStatus(e)}
+                onChange={(e) => handleStatusSelect(e)}
               >
                 <MenuItem value={'unread'}>Unread</MenuItem>
                 <MenuItem value={'read'}>Read</MenuItem>
@@ -248,7 +209,8 @@ export const BookForm = ({
                 <MenuItem value={'reading'}>Currently reading</MenuItem>
               </Select>
             </FormControl>
-            {showMeetingSelector && (
+            {(selectedReadStatus === 'read' ||
+              selectedReadStatus === 'reading') && (
               <FormControl>
                 <InputLabel id="meeting-select-label">
                   Select meeting
